risc0_ethereum_contracts
Enum SystemExitCode

#[repr(u8)]
pub enum SystemExitCode {
    Halted,
    Paused,
    SystemSplit,
    // some variants omitted
}

@notice Exit condition indicated by the zkVM at the end of the execution covered by this proof. @dev Halted indicates normal termination of a program with an interior exit code returned from the guest program. A halted program cannot be resumed.

Paused indicates the execution ended in a paused state with an interior exit code set by the guest program. A paused program can be resumed such that execution picks up where it left of, with the same memory state.

SystemSplit indicates the execution ended on a host-initiated system split. System split is mechanism by which the host can temporarily stop execution of the execution ended in a system split has no output and no conclusions can be drawn about whether the program will eventually halt. System split is used in continuations to split execution into individually provable segments.

enum SystemExitCode { Halted, Paused, SystemSplit }

Variants

Halted

Paused

SystemSplit